refactor: move EmptyList component  to shared folder
