+++ src/shared/board-list/index.ts
@@ -1,2 +1,3 @@
 export { BoardList } from './BoardList';
 export { PlaceHolder } from './PlaceHolder';
+export { EmptyList } from './EmptyList';
